Image Slice Converter Demo/requirements.txt
Pillow==10.2.0
PyQt5==5.15.10
pyinstaller==6.4.0